+++ src/main/scala/com/github/unchama/seichiassist/util/Util.scala
@@ -287,30 +287,6 @@ object Util {
     else AbsoluteDirection.NORTH
   }
 
-  def showTime(cal: Calendar): String = {
-    val date = cal.getTime
-    val format = new SimpleDateFormat("yyyy/MM/dd HH:mm")
-    format.format(date)
-  }
-
-  def showHour(cal: Calendar): String = {
-    val date = cal.getTime
-    val format = new SimpleDateFormat("HH:mm")
-    format.format(date)
-  }
-
-  def getTimeZone(cal: Calendar): String = {
-    val date = cal.getTime
-    val format = new SimpleDateFormat("HH")
-    val n = Integer.parseInt(format.format(date))
-    if (4 <= n && n < 10)
-      "morning"
-    else if (10 <= n && n < 18)
-      "day"
-    else
-      "night"
-  }
-
   def isVotingFairyPeriod(start: Calendar, end: Calendar): Boolean = {
     val cur = Calendar.getInstance()
     cur.after(start) && cur.before(end)
